+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [0.3.0-b.2] - 2021-01-21
+
+* Fix session level Flow frame handling
+
+* Cleanup unwraps
+
 ## [0.3.0-b.1] - 2021-01-19
 
 * Use ntex-0.2